Refactor location view
