+++ .eslintrc
@@ -4,6 +4,8 @@
   },
   "extends": "airbnb/base",
   "rules": {
+    "no-fallthrough": 0,
+    "no-case-declarations": 0,
     "no-console": 0,
     "no-plusplus": 0,
     "no-mutable-exports": 0,
