+++ src/common/pages/location/main_view.js
@@ -1,36 +1,18 @@
 /** ****************************************************************************
  * Location main view.
  *****************************************************************************/
-import $ from 'jquery';
 import Marionette from 'backbone.marionette';
-import L from 'leaflet';
-import 'leaflet/dist/leaflet.css';
-import 'leaflet/dist/images/layers-2x.png';
-import 'leaflet/dist/images/layers.png';
-import OSLeaflet from 'os-leaflet';
-import GridRef from 'leaflet.gridref';
-import { OsGridRef } from 'geodesy';
 import JST from 'JST';
-import { LocHelp, StringHelp } from 'helpers';
+import { LocHelp } from 'helpers';
 import CONFIG from 'config';
-import 'typeahead';
-import mapMarker from './map_view_marker';
-import locationNameFinder from './location_name_search';
-import LeafletButton from './leaflet_button_ext';
-
-const DEFAULT_LAYER = 'OS';
-const DEFAULT_CENTER = [53.7326306, -2.6546124];
-const MAX_OS_ZOOM = L.OSOpenSpace.RESOLUTIONS.length - 1;
-const OS_ZOOM_DIFF = 6;
-const OS_CRS = L.OSOpenSpace.getCRS(); // OS maps use different projection
-
-const GRID_STEP = 100000; // meters
-
+import 'typeahead'; //typeahead.js
+import headerFunctions from './main_view_header';
+import mapFunctions from './map/main';
 import './styles.scss';
 
 const LocationView = Marionette.View.extend({
   id: 'location-container',
-  template: JST['common/location/location'],
+  template: JST['common/location/main'],
 
   triggers: {
     'click #location-lock-btn': 'lock:click:location',
@@ -41,73 +23,10 @@ const LocationView = Marionette.View.extend({
   events: {
     'change #location-name': 'changeName',
     'typeahead:select #location-name': 'changeName',
-	  'change #location-gridref': 'changeGridRef',
+    'change #location-gridref': 'changeGridRef',
     'keyup #location-gridref': 'keyupGridRef',
     'blur #location-name': 'blurInput',
     'blur #location-gridref': 'blurInput',
-  },
-
-  changeName(e) {
-    this.triggerMethod('location:name:change', $(e.target).val());
-  },
-
-  blurInput() {
-    this._refreshMapHeight();
-  },
-
-  /**
-   * after delay, if gridref is valid then apply change
-   */
-  keyupGridRef(e) {
-    switch (e.keyCode) {
-      case 13:
-      // press Enter
-      case 38:
-      // Up
-      case 40:
-        // Down
-        break;
-      default:
-        // Other
-        let gr = e.target.value.replace(/\s+/g, '').toUpperCase();
-
-        if (gr === this._getCurrentLocation().gridref) {
-          return; // gridref hasn't changed meaningfully
-        }
-
-        // Clear previous timeout
-        this._clearGrTimeout();
-
-        if (gr === '' || LocHelp.grid2coord(gr)) {
-          // gr syntax ok (or blank)
-
-          this._refreshGrErrorState(false);
-
-          const that = this;
-          // Set new timeout - don't run if user is typing
-          this.grRefreshTimeout = setTimeout(function () {
-            // let controller know
-            that.trigger('location:gridref:change', gr);
-          }, 200);
-        } else {
-          this._refreshGrErrorState(true);
-        }
-    }
-  },
-
-  /**
-   * stop any delayed gridref refresh
-   */
-  _clearGrTimeout() {
-    if (this.grRefreshTimeout) {
-      clearTimeout(this.grRefreshTimeout);
-      this.grRefreshTimeout = null;
-    }
-  },
-
-  changeGridRef(e) {
-    this._clearGrTimeout();
-    this.triggerMethod('location:gridref:change', $(e.target).val());
   },
 
   initialize() {
@@ -120,7 +39,8 @@ const LocationView = Marionette.View.extend({
 
     const recordModel = this.model.get('recordModel');
 
-    //this.listenTo(recordModel, 'geolocation:start geolocation:stop geolocation:error', this.render);
+    // this.listenTo(recordModel,
+    // 'geolocation:start geolocation:stop geolocation:error', this.render);
     this.listenTo(recordModel, 'geolocation:start', this.geolocationStart);
     this.listenTo(recordModel, 'geolocation:stop', this.geolocationStop);
     this.listenTo(recordModel, 'geolocation:error', this.geolocationError);
@@ -138,426 +58,10 @@ const LocationView = Marionette.View.extend({
     this.addLocationNameSearch();
   },
 
-  /**
-   * set full remaining height
-   *
-   */
-  _refreshMapHeight() {
-    //const mapHeight = $(document).height() - 47 - 47 - 44;// - 47 - 38.5;
-    this.$container = this.$el.find('#map')[0];
-    //$(this.$container).height(mapHeight);
-    $(this.$container).style = "height: 100vh;";
-  },
-
-  addLocationNameSearch() {
-    this.$el.find('.typeahead').typeahead({
-      hint: false,
-      highlight: false,
-      minLength: 0,
-    },
-      {
-        limit: 3,
-        name: 'names',
-        source: locationNameFinder(3),
-      });
-  },
-
-  initMap() {
-    this.map = L.map(this.$container);
-
-    // default layer
-    this.currentLayer = this._getCurrentLayer();
-    if (this.currentLayer === 'OS') this.map.options.crs = OS_CRS;
-
-    // position view
-    this.map.setView(this._getCenter(), this._getZoomLevel());
-
-    // show default layer
-    this.layers[this.currentLayer].addTo(this.map);
-    this.$container.dataset.layer = this.currentLayer; // fix the lines between the tiles
-
-    this.map.on('baselayerchange', this._updateCoordSystem, this);
-    this.map.on('zoomend', this.onMapZoom, this);
-
-    // Controls
-    this.addControls();
-
-    // GPS
-    this.addGPS();
-
-    // Marker
-    this.addMapMarker();
-
-    // Graticule
-    this.addGraticule();
-  },
-
-  _getLayers() {
-    const layers = {};
-    layers.Satellite = L.tileLayer('https://api.tiles.mapbox.com/v4/{id}/{z}/{x}/{y}.png?access_token={accessToken}', {
-      attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://mapbox.com">Mapbox</a>',
-      id: CONFIG.map.mapbox_satellite_id,
-      accessToken: CONFIG.map.mapbox_api_key,
-      tileSize: 256, // specify as, OS layer overwites this with 200 otherwise,
-      minZoom: 5,
-    });
-
-    layers.OSM = L.tileLayer('https://api.tiles.mapbox.com/v4/{id}/{z}/{x}/{y}.png?access_token={accessToken}', {
-      attribution: 'Map data &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors, <a href="http://creativecommons.org/licenses/by-sa/2.0/">CC-BY-SA</a>, Imagery © <a href="http://mapbox.com">Mapbox</a>',
-      id: CONFIG.map.mapbox_osm_id,
-      accessToken: CONFIG.map.mapbox_api_key,
-      tileSize: 256, // specify as, OS layer overwites this with 200 otherwise
-      minZoom: 5,
-    });
-
-    const start = OsGridRef.osGridToLatLon(OsGridRef(0, 0));
-    const end = OsGridRef.osGridToLatLon(OsGridRef(7 * GRID_STEP, 13 * GRID_STEP));
-    const bounds = L.latLngBounds([start.lat, start.lon], [end.lat, end.lon]);
-
-    layers.OS = L.tileLayer.OSOpenSpace(CONFIG.map.os_api_key);
-
-    layers.OS.options.bounds = bounds;
-
-    layers.OS.on('tileerror', (tile) => {
-      let index = 0;
-      const result = tile.tile.src.match(/missingTileString=(\d+)/i);
-      if (result) {
-        index = parseInt(result[1], 10);
-        index++;
-
-        // don't do it more than few times
-        if (index < 4) {
-          tile.tile.src = tile.tile.src.replace(/missingTileString=(\d+)/i, '&missingTileString=' + index);
-        }
-      } else {
-        if (index === 0) {
-          tile.tile.src = tile.tile.src + '&missingTileString=' + index;
-        }
-      }
-    });
-    return layers;
-  },
-
-  _getCurrentLayer() {
-    let layer = DEFAULT_LAYER;
-    const zoom = this._getZoomLevel();
-    const currentLocation = this._getCurrentLocation();
-    const inUK = LocHelp.isInUK(currentLocation);
-    if (zoom > MAX_OS_ZOOM - 1) {
-      layer = 'Satellite';
-    } else if (inUK === false) {
-      this.currentLayerControlSelected = true;
-      layer = 'Satellite';
-    }
-
-    return layer;
-  },
-
-  _getCenter() {
-    const currentLocation = this._getCurrentLocation();
-    let center = DEFAULT_CENTER;
-    if (currentLocation.latitude) {
-      center = [currentLocation.latitude, currentLocation.longitude];
-    }
-    return center;
-  },
-
-  addControls() {
-    this.controls = L.control.layers({
-      OS: this.layers.OS,
-      OSM: this.layers.OSM,
-      Satellite: this.layers.Satellite,
-    }, {});
-    this.map.addControl(this.controls);
-  },
-
-  addGPS() {
-    const that = this;
-    const location = this._getCurrentLocation();
-
-    const GPSbutton = new LeafletButton({
-      position: 'topright',
-      className: 'gps-btn',
-      body: `<span class="icon icon-location"
-                data-source="${location.source}"></span>`,
-      'onClick'() {
-        that.trigger('gps:click');
-      },
-      'maxWidth': 30,  // number
-    });
-
-
-    this.map.addControl(GPSbutton);
-    const recordModel = this.model.get('recordModel');
-    if (recordModel.isGPSRunning()) {
-      this._set_gps_progress_feedback('pending');
-    } else {
-      this._set_gps_progress_feedback('');
-    }
-  },
-
-  addGraticule() {
-    const appModel = this.model.get('appModel');
-    const useGridRef = appModel.get('useGridRef');
-    const useGridMap = appModel.get('useGridMap');
-    if (!useGridRef || !useGridMap) return;
-
-    const that = this;
-
-    function getColor() {
-      'use strict';
-      let color;
-      switch (that.currentLayer) {
-        case 'OS':
-          color = '#08b7e8';
-          break;
-        case 'OSM':
-          color = 'gray';
-          break;
-        default:
-          color = 'white';
-      }
-      return color;
-    }
-
-    const gridRef = new L.GridRef({ color: getColor() });
-
-    gridRef.update = () => {
-      let zoom = that.map.getZoom();
-      // calculate granularity
-      const color = getColor();
-      if (that.currentLayer === 'OS') zoom += OS_ZOOM_DIFF;
-
-      const bounds = that.map.getBounds();
-      const granularity = gridRef._getGranularity(zoom);
-      const step = GRID_STEP / granularity;
-
-      const polylinePoints = gridRef._calcGraticule(step, bounds);
-      gridRef.setStyle({ color });
-      gridRef.setLatLngs(polylinePoints);
-    };
-    gridRef.addTo(this.map);
-  },
-
-  /**
-   * 1 gridref digits. (10000m)  -> < 3 map zoom lvl
-   * 2 gridref digits. (1000m)   -> 5
-   * 3 gridref digits. (100m)    -> 7
-   * 4 gridref digits. (10m)     -> 9
-   * 5 gridref digits. (1m)      ->
-   */
-  _getZoomLevel() {
-    const currentLocation = this._getCurrentLocation();
-    let mapZoomLevel = 1;
-
-    // check if record has location
-    if (currentLocation.latitude) {
-      // transform location accuracy to map zoom level
-
-      /**
-        * 1 gridref digits. (10000m)  -> 4 OS map zoom lvl
-        * 2 gridref digits. (1000m)   -> 8 OS
-        * 3 gridref digits. (100m)    -> 16 OSM
-        * 4 gridref digits. (10m)     -> 18 OSM
-        * 5 gridref digits. (1m)      ->
-        */
-      if (currentLocation.accuracy) {
-        if (currentLocation.accuracy > 1000) {
-          mapZoomLevel = 4;
-        } else if (currentLocation.accuracy > 100) {
-          mapZoomLevel = 8;
-        } else if (currentLocation.accuracy > 10) {
-          mapZoomLevel = 16;
-        } else {
-          mapZoomLevel = 18;
-        }
-      } else {
-        mapZoomLevel = 1;
-      }
-    }
-    //if (this.currentLayer && this.currentLayer !== 'OS' && mapZoomLevel < MAX_OS_ZOOM) {
-    if (this.currentLayer && this.currentLayer !== 'OS') {
-      mapZoomLevel += OS_ZOOM_DIFF;
-
-      if (mapZoomLevel > 18) {
-        mapZoomLevel = 18;
-      }
-    }
-
-    return mapZoomLevel;
-  },
-
-  _updateCoordSystem(e) {
-    this.currentLayerControlSelected = this.controls._handlingClick;
-
-    const center = this.map.getCenter();
-    let zoom = this.map.getZoom();
-    this.map.options.crs = e.name === 'OS' ? OS_CRS : L.CRS.EPSG3857;
-
-    if (!this.noZoomCompensation) {
-      if (e.name === 'OS') {
-        zoom -= OS_ZOOM_DIFF;
-
-        if (zoom > MAX_OS_ZOOM - 1) {
-          zoom = MAX_OS_ZOOM - 1;
-        }
-      } else if (this.currentLayer === 'OS') {
-        zoom += OS_ZOOM_DIFF;
-      }
-    }
-    this.currentLayer = e.name;
-    this.map.setView(center, zoom, { reset: true });
-    this.$container.dataset.layer = this.currentLayer; // fix the lines between the tiles
-  },
-
-  onMapZoom() {
-    const zoom = this.map.getZoom();
-    const inUK = LocHelp.isInUK(this._getCurrentLocation());
-
-    // -2 and not -1 because we ignore the last OS zoom level
-    if (zoom > MAX_OS_ZOOM - 1 && this.currentLayer === 'OS') {
-      this.map.removeLayer(this.layers.OS);
-      this.map.addLayer(this.layers.Satellite);
-    } else if ((zoom - OS_ZOOM_DIFF) <= MAX_OS_ZOOM - 1 && this.currentLayer === 'Satellite') {
-      // only change base layer if user is on OS and did not specificly
-      // select OSM/Satellite
-      if (!this.currentLayerControlSelected && inUK !== false) {
-        this.map.removeLayer(this.layers.Satellite);
-        this.map.addLayer(this.layers.OS);
-      }
-    }
-  },
-
-  geolocationStart() {
-    this._set_gps_progress_feedback('pending');
-  },
-
-  /**
-   * Update the temporary location fix
-   * @param location
-   */
-  geolocationUpdate(location) {
-    this.locationUpdate = location;
-    this._set_gps_progress_feedback('pending');
-  },
-
-  geolocationSuccess(location) {
-    this.locationUpdate = location;
-    this._set_gps_progress_feedback('fixed');
-  },
-
-  geolocationStop() {
-    this._set_gps_progress_feedback('');
-  },
-
-  geolocationError() {
-    this._set_gps_progress_feedback('failed');
-  },
-
-  _set_gps_progress_feedback(state) {
-    const $gpsButton = this.$el.find('.gps-btn');
-    // change state
-    $gpsButton.attr('data-gps-progress', state);
-
-
-    // change icon
-    const $gpsButtonSpan = $gpsButton.find('span');
-    if (state === 'pending') {
-      $gpsButtonSpan.addClass('icon-plus icon-spin');
-      $gpsButtonSpan.removeClass('icon-location');
-
-    } else {
-      $gpsButtonSpan.removeClass('icon-plus icon-spin');
-      $gpsButtonSpan.addClass('icon-location');
-    }
-
-  },
-
-  _refreshGrErrorState(isError) {
-    const grInputEl = document.getElementById('location-gridref');
-    if (grInputEl) {
-      if (isError) {
-        grInputEl.setAttribute('data-gr-error', 'error');
-        this._removeMapMarker();
-      } else {
-        grInputEl.removeAttribute('data-gr-error');
-      }
-    }
-  },
-
-  locationChange() {
-    this._clearGrTimeout();
-    const location = this._getCurrentLocation();
-
-    this._refreshGrErrorState(false);
-
-    this.updateMapMarker(location);
-
-    //if source was 'map' then presume that current zoom is fine so don't change (send undefined)
-    //this.map.setView(this._getCenter(), location.source !== 'map' ? this._getZoomLevel() : undefined);
-    this.noZoomCompensation = false;
-    //this.map.setView(this._getCenter(), this._getZoomLevel());
-    this.map.setView(this._getCenter(), location.source !== 'map' ? this._getZoomLevel() : undefined);
-    //this.map.panTo(this._getCenter());
-    //this.map.setZoom(this._getZoomLevel());
-    this.noZoomCompensation = false;
-    this._refreshGridRefElement(location);
-  },
-
-  _getCurrentLocation() {
-    return this.model.get('recordModel').get('location') || {};
-  },
-
-  _refreshGridRefElement(location) {
-    // rather than full refresh of the view, directly update the relavant input element
-    const $GR = this.$el.find('#location-gridref');
-    $GR.val(location.gridref);
-    $GR.attr('data-source', location.source);
-
-    const $gpsBtn = this.$el.find('.gps-btn');
-    if ($gpsBtn) {
-      $gpsBtn.attr('data-source', location.source);
-      
-      if (location.source !== 'gps') {
-        this._set_gps_progress_feedback('');
-      }
-    }
-  },
-
-  updateLocks() {
-    const appModel = this.model.get('appModel');
-    const recordModel = this.model.get('recordModel');
-    const location = recordModel.get('location') || {};
-    const location_name = recordModel.get('location_name');
-
-    // location lock
-    const $locationLockBtn = this.$el.find('#location-lock-btn');
-    const locationLocked = appModel.isAttrLocked('location', location);
-    if (locationLocked) {
-      $locationLockBtn.addClass('icon-lock-closed');
-      $locationLockBtn.removeClass('icon-lock-open');
-    } else {
-      $locationLockBtn.addClass('icon-lock-open');
-      $locationLockBtn.removeClass('icon-lock-closed');
-    }
-
-    // location name lock
-    const $nameLockBtn = this.$el.find('#name-lock-btn');
-    const nameLocked = appModel.isAttrLocked('location_name', location_name);
-    if (nameLocked) {
-      $nameLockBtn.addClass('icon-lock-closed');
-      $nameLockBtn.removeClass('icon-lock-open');
-    } else {
-      $nameLockBtn.addClass('icon-lock-open');
-      $nameLockBtn.removeClass('icon-lock-closed');
-    }
-
-  },
-
   serializeData() {
     const appModel = this.model.get('appModel');
     const location = this._getCurrentLocation();
-    const location_name = this.model.get('recordModel').get('location_name')
+    const name = this.model.get('recordModel').get('location_name');
     let gridref;
 
     // avoid testing location.longitude as this can validly be zero within the UK
@@ -568,11 +72,11 @@ const LocationView = Marionette.View.extend({
     }
 
     const locationLocked = appModel.isAttrLocked('location', location);
-    const nameLocked = appModel.isAttrLocked('location_name', location_name);
+    const nameLocked = appModel.isAttrLocked('location_name', name);
 
     return {
-      name: location_name,
-      gridref: gridref,
+      name,
+      gridref,
       locationSource: location.source,
       accuracy: location.accuracy,
       latitude: location.latitude,
@@ -584,4 +88,5 @@ const LocationView = Marionette.View.extend({
   },
 });
 
-export default LocationView.extend(mapMarker);
+const LocationViewHeader = LocationView.extend(headerFunctions);
+export default LocationViewHeader.extend(mapFunctions);
