+++ src/common/pages/location/map/leaflet_singleclick_ext.js
