+++ src/common/pages/location/styles.scss
@@ -1,5 +1,4 @@
 @import "../../styles/_variables.scss";
-@import "./leaflet_btn.scss";
 
 #map-location-name {
   position: absolute;
