--- src/common/pages/location/templates/gps_error.tpl
@@ -1,4 +1,0 @@
-<div class="error-message">
-  <p>Sorry!</p>
-  <p>We could not find your location.</p>
-</div>
