--- src/common/pages/location/templates/gps_success.tpl
@@ -1,13 +1,0 @@
-<div class="success-message">
-  <p>Success!</p>
-  <p>Accuracy: <%- obj.accuracy %> meters</p>
-</div>
-
-<div class="input-row tt">
-  <label class="media-object pull-left icon icon-address"></label>
-  <input class="typeahead" type="text" id="location-name" placeholder="Nearest Named Place" value="<%= obj.name %>"/>
-</div>
-
-<button id="gps-button"
-        class="btn btn-narrow btn-positive btn-block">Refresh GPS</button>
-
