+++ src/common/pages/location/templates/main.tpl
