--- src/common/pages/location/templates/past.tpl
@@ -1,4 +1,0 @@
-<div class="info-message" id="previous-location-message">
-  <p>Please tap on your previous location. </p>
-</div>
-<div id="locations"></div>
